rename project_issue_correlation to get_suspect_resolutions
